requirements.txt
google-generativeai
Pillow
pyautogui
plyer
matplotlib
requests
pygetwindow; sys_platform == "win32"
